feat: authorize claimer & cleanup
