+++ src/components/delegate/delegate-section.tsx
@@ -1,0 +1,105 @@
+import { RVR_AUTHORIZER, RVR_TOKEN } from "@/constants/contracts"
+import { cn } from "@/lib/utils"
+import { useAccount, useReadContract, useReadContracts } from "wagmi"
+import { Typography } from "../ui/typography"
+import { WalletAddress } from "../wallet-address"
+import { formatUnits } from "viem"
+import { Skeleton } from "../ui/skeleton"
+import { DelegateForm } from "./delegate-form"
+import { AuthorizeClaimerForm } from "./authorize-claimer-form"
+
+export const DelegateSection = () => {
+  const { address } = useAccount()
+
+  const riverToken = useReadContracts({
+    allowFailure: false,
+    contracts: [
+      {
+        address: RVR_TOKEN.sepolia,
+        abi: RVR_TOKEN.abi,
+        functionName: 'balanceOf',
+        args: address ? [address] : undefined,
+      },
+      {
+        address: RVR_TOKEN.sepolia,
+        abi: RVR_TOKEN.abi,
+        functionName: 'decimals',
+      },
+    ],
+    query: {
+      enabled: !!address,
+    },
+  })
+
+  const delegatee = useReadContract({
+    address: RVR_TOKEN.sepolia,
+    abi: RVR_TOKEN.abi,
+    functionName: 'delegates',
+    args: address ? [address] : undefined,
+  })
+
+  const authorizedClaimer = useReadContract({
+    address: RVR_AUTHORIZER.sepolia,
+    abi: RVR_AUTHORIZER.abi,
+    functionName: 'getAuthorizedClaimer',
+    args: address ? [address] : undefined,
+  })
+
+  return (
+    <section
+      className={cn(
+        'hero-glow relative w-full overflow-x-clip bg-gray-90 md:min-h-screen',
+        'flex w-full flex-col items-center',
+        'pb-24 pt-[88px]',
+      )}
+    >
+      <section className={cn('flex w-full flex-col gap-6 px-4', 'sm:max-w-lg')}>
+        <Typography
+          as="h1"
+          size="3xl"
+          className="text-2xl font-semibold leading-[44px] text-gray-10"
+        >
+          Delegate
+        </Typography>
+
+        <div className="w-full rounded-3xl border border-solid border-gray-60 bg-gray-80 p-6">
+          <div className="flex justify-between gap-4">
+            <span className="text-gray-20">Connected:</span>
+            {address && <WalletAddress address={address} />}
+          </div>
+          <div className="flex justify-between">
+            <span className="text-gray-20">RVR Balance:</span>
+            <span className="text-white">
+              {riverToken.isLoading ? (
+                <Skeleton className="inline-block h-4 w-20" />
+              ) : riverToken.data?.[0] && riverToken.data?.[1] ? (
+                formatUnits(riverToken.data[0], riverToken.data[1])
+              ) : (
+                0
+              )}
+            </span>
+          </div>
+          <div className="flex justify-between">
+            <span className="text-gray-20">Delegating to:</span>
+            {delegatee.isLoading ? (
+              <Skeleton className="inline-block h-4 w-36" />
+            ) : delegatee.data ? (
+              <WalletAddress address={delegatee.data} />
+            ) : null}
+          </div>
+          <div className="flex justify-between">
+            <span className="text-gray-20">Authorized claimer:</span>
+            {authorizedClaimer.isLoading ? (
+              <Skeleton className="inline-block h-4 w-36" />
+            ) : !authorizedClaimer.data ? null : (
+              <WalletAddress address={authorizedClaimer.data} />
+            )}
+          </div>
+        </div>
+
+        <DelegateForm delegateeQueryKey={delegatee.queryKey} />
+        <AuthorizeClaimerForm authorizedClaimerQueryKey={authorizedClaimer.queryKey} />
+      </section>
+    </section>
+  )
+}
