+++ src/lib/wagmi.ts
@@ -1,4 +1,5 @@
 import { defaultWagmiConfig } from '@web3modal/wagmi/react/config'
+import { cookieStorage, createStorage } from 'wagmi'
 
 import { mainnet, sepolia } from 'wagmi/chains'
 
@@ -19,4 +20,7 @@ export const wagmiConfig = defaultWagmiConfig({
   projectId,
   metadata,
   ssr: true,
+  storage: createStorage({  
+    storage: cookieStorage, 
+  }), 
 })
